Use best-effort rather than failing, when no feasible solutions
Now, `addFeasible` and `filterInfeasible` will never let the list of
active breakpoints be empty. This means that a layout will always be
produced, even if it's outside the typical range of allowed adjustment
ratios.

But, if an optimal layout is possible, it should still always be
chosen.

Also, some additional debug info was added to the output display (the
index of each fragment, and the adjustment ratio of each line).